feat: Migrate `IncognitoScreen` and `IncognitoViewModel` to kmpuiviews
This commit moves the `IncognitoScreen` and `IncognitoViewModel` to the `kmpuiviews` module, enabling them to be used in multiplatform projects.

- **IncognitoScreen and IncognitoViewModel:**
    -   Moved `IncognitoScreen` from `UIViews` to `kmpuiviews`.
    -   Moved `IncognitoViewModel` from `UIViews` to `kmpuiviews`.
    -   Added `IncognitoModel` to `kmpuiviews`.
    -   Removed the old `IncognitoScreen` and `IncognitoViewModel` from `UIViews`.
    - Updated `NavGraph` in `UIViews` to remove the old `IncognitoScreen`
    - Updated `ViewModelModule` to include `kmpuiviews` view models.
-   - Removed the `IncognitoViewModel` from `ViewModelModule` in `UIViews`.
- **Dependencies:**
-    - Removed unnecessary dependencies.
- **GradientImage:**
+   - Updated `ViewModelModule` in `UIViews` to remove `IncognitoViewModel`.
+- **GradientImage:**
     - Moved `GradientImage` from `UIViews` to `kmpuiviews` and removed the old version.
     - Updated `NotificationFragment` and `HistoryFragment` to use the new `GradientImage`.
-  - Updated `CoilGradientImage` to `GradientImage`
+- **General:**
+    - Updated `CoilGradientImage` to `GradientImage`.
+    - Created `ViewModelModule` in `kmpuiviews`.
