+++ UIViews/src/main/java/com/programmersbox/uiviews/di/ViewModelModule.kt
@@ -1,6 +1,7 @@
 package com.programmersbox.uiviews.di
 
 import com.programmersbox.datastore.DataStoreHandling
+import com.programmersbox.kmpuiviews.di.viewModels
 import com.programmersbox.uiviews.presentation.all.AllViewModel
 import com.programmersbox.uiviews.presentation.details.DetailsViewModel
 import com.programmersbox.uiviews.presentation.favorite.FavoriteViewModel
@@ -14,7 +15,6 @@ import com.programmersbox.uiviews.presentation.notifications.NotificationScreenV
 import com.programmersbox.uiviews.presentation.recent.RecentViewModel
 import com.programmersbox.uiviews.presentation.settings.downloadstate.DownloadStateViewModel
 import com.programmersbox.uiviews.presentation.settings.extensions.ExtensionListViewModel
-import com.programmersbox.uiviews.presentation.settings.extensions.IncognitoViewModel
 import com.programmersbox.uiviews.presentation.settings.moresettings.MoreSettingsViewModel
 import com.programmersbox.uiviews.presentation.settings.updateprerelease.PrereleaseViewModel
 import com.programmersbox.uiviews.presentation.settings.viewmodels.AccountViewModel
@@ -45,5 +45,5 @@ fun Module.viewModels() {
     viewModelOf(::PrereleaseViewModel)
     viewModelOf(::DownloadStateViewModel)
     viewModelOf(::AccountViewModel)
-    viewModelOf(::IncognitoViewModel)
+    includes(viewModels)
 }
