+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/NavGraph.kt
@@ -20,6 +20,8 @@ import androidx.navigation.navigation
 import androidx.navigation.toRoute
 import com.programmersbox.gemini.GeminiRecommendationScreen
 import com.programmersbox.kmpuiviews.presentation.Screen
+import com.programmersbox.kmpuiviews.presentation.settings.incognito.IncognitoScreen
+import com.programmersbox.kmpuiviews.presentation.webview.WebViewScreen
 import com.programmersbox.kmpuiviews.utils.ComposeSettingsDsl
 import com.programmersbox.kmpuiviews.utils.chromeCustomTabs
 import com.programmersbox.uiviews.BuildConfig
@@ -44,10 +46,8 @@ import com.programmersbox.uiviews.presentation.settings.SettingScreen
 import com.programmersbox.uiviews.presentation.settings.SourceOrderScreen
 import com.programmersbox.uiviews.presentation.settings.downloadstate.DownloadStateScreen
 import com.programmersbox.uiviews.presentation.settings.extensions.ExtensionList
-import com.programmersbox.uiviews.presentation.settings.extensions.IncognitoScreen
 import com.programmersbox.uiviews.presentation.settings.moresettings.MoreSettingsScreen
 import com.programmersbox.uiviews.presentation.settings.updateprerelease.PrereleaseScreen
-import com.programmersbox.kmpuiviews.presentation.webview.WebViewScreen
 import com.programmersbox.uiviews.utils.NotificationLogo
 import com.programmersbox.uiviews.utils.sharedelements.animatedScopeComposable
 import com.programmersbox.uiviews.utils.trackScreen
