+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/components/GradientImage.kt
@@ -13,7 +13,6 @@ import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.ColorMatrix
 import androidx.compose.ui.graphics.TileMode
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
@@ -23,56 +22,6 @@ import com.bumptech.glide.integration.compose.CrossFade
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.bumptech.glide.integration.compose.placeholder
-import io.kamel.core.Resource
-import io.kamel.image.KamelImage
-import io.kamel.image.asyncPainterResource
-
-@Composable
-fun GradientImage(
-    model: Any,
-    placeholder: Painter,
-    modifier: Modifier = Modifier,
-    error: Painter = placeholder,
-    contentDescription: String? = null,
-    contentScale: ContentScale = ContentScale.FillBounds,
-    blur: Dp = 70.dp,
-    alpha: Float = .5f,
-    saturation: Float = 3f,
-    scaleX: Float = 1.5f,
-    scaleY: Float = 1.5f,
-) {
-    Box {
-        when (val painter = asyncPainterResource(data = model)) {
-            is Resource.Success -> {
-                KamelImage(
-                    resource = { painter },
-                    contentDescription = contentDescription,
-                    contentScale = contentScale,
-                    colorFilter = ColorFilter.colorMatrix(ColorMatrix().apply { setToSaturation(saturation) }),
-                    modifier = Modifier
-                        .blurGradient(blur, alpha, scaleX, scaleY)
-                        .then(modifier)
-                )
-
-                KamelImage(
-                    resource = { painter },
-                    contentDescription = contentDescription,
-                    contentScale = contentScale,
-                    modifier = modifier
-                )
-            }
-
-            else -> {
-                Image(
-                    painter = error,
-                    contentScale = contentScale,
-                    contentDescription = contentDescription,
-                    modifier = modifier
-                )
-            }
-        }
-    }
-}
 
 @Composable
 fun CoilGradientImage(
