+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/di/ViewModelModule.kt
@@ -1,0 +1,10 @@
+package com.programmersbox.kmpuiviews.di
+
+import com.programmersbox.kmpuiviews.presentation.settings.incognito.IncognitoViewModel
+import org.koin.core.module.Module
+import org.koin.core.module.dsl.viewModelOf
+import org.koin.dsl.module
+
+val viewModels: Module = module {
+    viewModelOf(::IncognitoViewModel)
+}
